feat(templates): Library search/filters; split view into components
Library tab gets the same toolbar as Bindings — search by name or
used-by key, source select, All/Used/Unused segmented. Mechanical
decomposition of Templates.tsx (1119 → 105 lines) into
views/templates/{helpers,KeyBuilder,TemplatePicker,RowEditorModal,
ImportWizard,BindingsTable,InventoryPanel}.
